create a name.rb file with name non-SearchWorks name related methods
